extract code from extension.ts
