+++ extension/src/extension.ts
@@ -2,19 +2,15 @@ import * as vscode from "vscode";
 import * as path from "path";
 import { ComponentViewerProvider } from "./componentViewerProvider";
 import { CompositionViewerProvider } from "./compositionViewerProvider";
-import { exec, execSync } from "child_process";
 import * as promisify from "es6-promisify";
-import * as portscanner from "portscanner";
-import * as opn from "opn";
 import * as freeport from "freeport";
 import { OutputChannel } from "vscode";
 import * as fs from "fs";
-import { parseStringSync } from "xml2js-parser";
-import { getServerUrl } from "./webSocketConfigParser";
+import { build } from "./projectBuilder";
+import { spyExec } from "./spyExec";
 
 const xcmlExtension = ".xcml";
 const cxmlExtension = ".cxml";
-const openBrowserTimeOut = 2000;
 
 interface DataUrl {
     componentName: string;
@@ -67,58 +63,13 @@ export function activate(context: vscode.ExtensionContext) {
     });
 
     const disposableBuild = vscode.commands.registerCommand("xcomponent.build.project", () => {
-        const xcbuildPath = vscode.workspace.getConfiguration()["xcbuild"];
-        if (!fs.existsSync(xcbuildPath)) {
-            vscode.window.showErrorMessage(`xcbuild.exe not found. Please specify xcbuild path in vscode settings`);
-            return;
-        }
-        const rootPath = vscode.workspace.rootPath;
-        const baseName = path.basename(rootPath);
-        const xcmlPath = `${rootPath}${path.sep}${baseName}_Model.xcml`;
-        if (!fs.existsSync(xcmlPath)) {
-            vscode.window.showErrorMessage(`Build error. File ${xcmlPath} not found`);
-            return;
-        }
-        const isWindowsPlatform = /^win/.test(process.platform);
-        if (isWindowsPlatform) {
-            const buildCommand = `${xcbuildPath} --compilationmode=Debug --build --env=Dev --vs=VS2015 --project=${xcmlPath}`;
-            terminal.show();
-            terminal.sendText(buildCommand);
-            return;
-        }
-        const monoPath = vscode.workspace.getConfiguration()["mono"];
-        if (!fs.existsSync(monoPath)) {
-            vscode.window.showErrorMessage(`mono not found. Please specify mono path in vscode settings`);
-            return;
-        }
-        const buildCommand = `mono ${xcbuildPath}--compilationmode=Release --build --framework=Framework452 --env=Dev --vs=VS2015 --monoPath=“${monoPath}” --project=“${xcmlPath}”`
-        terminal.show();
-        terminal.sendText(buildCommand);
+        build(terminal);
     });
-
 
     promisify(freeport)()
         .then((port: number) => {
             const disposableSpy = vscode.commands.registerCommand("xcomponent.launch.spy", () => {
-                const rootPath = vscode.workspace.rootPath;
-                const baseName = path.basename(rootPath);
-                const configurationFilePath = `${rootPath}${path.sep}Configuration.${baseName}${path.sep}Dev${path.sep}${baseName}_Deployment_Configuration.xml`;
-                const serverUrl = getServerUrl(configurationFilePath);
-                const binPath = path.join(context.extensionPath, "out/spy/bin");
-                const runSpyServercommand = `serve --single --port ${port} ${binPath}`;
-                const promiseCheckPortStatus = promisify(portscanner.checkPortStatus);
-                const urlParams = (serverUrl === undefined) ? "" : `/form?serverUrl=${serverUrl}`;
-                const url = `http://localhost:${port}${urlParams}`;
-                promiseCheckPortStatus(port, "localhost")
-                    .then((status: string) => {
-                        if (status === "closed") {
-                            terminal.show();
-                            terminal.sendText(runSpyServercommand);
-                            setTimeout(() => opn(url), openBrowserTimeOut);
-                        } else if (status === "open") {
-                            setTimeout(() => opn(url), openBrowserTimeOut);
-                        }
-                    });
+                spyExec(terminal, port, context.extensionPath);
             });
             context.subscriptions.push(disposableSpy);
         })
